double instead of unsigned

+++ image.h
@@ -11,16 +11,16 @@
 
 // Specify filepath of image to be read, width/height of image, and a 2D vec of unsigned ints (any size).  Modifies provided 2D vector to contain pixels represented as unsigned ints.
 // Note: This is designed only for colored, RGB PNG images with 3 bytes per pixel (hence width of 2D vec is width*3).
-void image_read(char const *filepath, int width, int height, vector<vector<unsigned> > &vec) {
+void image_read(char const *filepath, int width, int height, vector<vector<double> > &vec) {
     int bpp = 3;
     uint8_t* rgb_image = stbi_load(filepath, &width, &height, &bpp, 3);
     vec.resize(height);
     for (int i = 0; i < vec.size(); ++i) {
-        vec[i] = vector<unsigned>(width*bpp,0);
+        vec[i] = vector<double>(width*bpp,0);
     }
     for (int i = 0; i < height; ++i) {
         for (int j = 0; j < width*bpp; ++j) {
-            vec[i][j] = unsigned(rgb_image[width*bpp*i+j]);
+            vec[i][j] = double(rgb_image[width*bpp*i+j]);
         }
     }
     stbi_image_free(rgb_image);
@@ -28,7 +28,7 @@ void image_read(char const *filepath, int width, int height, vector<vector<unsig
 
 // Specify filepath where image should be written, and provide 2D vec of unsigned ints containing pixels of image.
 // Note: This is designed only for colored, RGB PNG images with 3 bytes per pixel
-void image_write(char const *filepath, vector<vector<unsigned> > &vec) {
+void image_write(char const *filepath, vector<vector<double> > &vec) {
     int width = vec[0].size();
     int height = vec.size();
     int bpp = 3;
